requirements.txt
# 프로젝트 의존성

yt-dlp>=2025.6.1
git+https://github.com/openai/whisper.git@main
openai>=0.27.0
torch>=2.0.0
google-generativeai>=0.3.0
demucs>=4.0.0

# Visualization & audio analysis
matplotlib>=3.8.0
librosa>=0.10.1
soundfile>=0.12.1
numpy>=1.24.0
pydub>=0.25.1
